MinMax now works with all supported data types.
Remrkably, uint8 was propagated to opencl as char. It was necessary to
use uchar for uint8 for this to work.  I realized that there may be
many cases where the type conversion are not happening correctly.
It would be good to state upfront which types are supported and to
ensure that everything works for those types.