Convenient histogram function added.
Also discovered that using ClearCLImageInterface rather than ClearCLImage
and Buffer can reduce the size of Kernel.java in half.  However, that needs
testing of each converted function.

Note: the kernel code for histograms assumes 256 bins integers.  It also
uses floats as the histogram data type.  It would be much better and
nicer to leave the number of bins up to the caller and to use an
integer data type.